+++ docs/changelog.markdown
@@ -9,6 +9,10 @@ nav_order: 100
 
 To read all the details go to the [Github Releases section](https://github.com/innoveit/react-native-ble-manager/releases).
 
+## Unreleased
+
+- [Android only] Added `useScanIntent` scan option to receive results via `PendingIntent` on API 26+.
+
 ## Release v12.1.X
 
 - Added expo plugin.